apusys power: fix debug message and add new governor
[Detail]
fix debug message and add new governor

(cherry-picked from commit da88ff3e93c6463a70473a4e6af73f9b78f7f3ba)
CR-Id: ALPS05487649
Feature: [Module] APUSYS
(cherry picked from commit f2a7cd1681bc6232ecf2f8c9ece4a45505fb1c63)

+++ drivers/misc/mediatek/apusys/power/2.5/inc/apu_of.h
@@ -50,4 +50,4 @@ int of_apu_link(struct device *dev, struct device_node *con_np, struct device_no
 
 void of_apu_regulator_put(struct apu_regulator *rgul);
 int of_apu_regulator_get(struct device *dev,
-		struct apu_regulator *rgul, int volt, ulong freq);
+		struct apu_regulator *rgul, unsigned long volt, ulong freq);
